wit, cmd/wit-bindgen-go: WIT tree shaking
This enables WIT tree-shaking by world or interface, as a precursor for Go package-level component metadata.

wit/bindgen: oops

wit/bindgen: add WIT generation for each WIT interface (Go package)

Currently disabled.

wit/bindgen: cgo + linker tricks WIP

wit: fix typo

wit/bindgen: typo

wit: type aliases force transitive dependency on the dependencies of their parent interface

wit: prepare for filtering interface contents

wit: remove ConstrainTo

internal/wasm: stub linking section

wit/bindgen: generate .wasm.syso files in each Go package

Currently stubbed out (if false). Depends on https://github.com/tinygo-org/tinygo/pull/4593

wit: revise package sorting algorithm

This enables wasi:http to sort before wasi:cli.

wit/bindgen: more cleanup (path -> pkgPath)

wit/bindgen: optionally generate WIT files for each Go package

cmd/wit-bindgen-go: --generate-wit option to generate WIT files for each Go package

wit/bindgen: generate synthetic worlds in the go:bindgen package namespace
